An enormous amount of outlook tweaks and problem fixes concerning individual characters inter alia.

+++ Main/Source/worldmap.cpp
@@ -25,7 +25,7 @@ worldmap::worldmap(ushort XSize, ushort YSize) : area(XSize, YSize)
   Map = (wsquare***)area::Map;
 
   for(ushort x = 0; x < XSize; ++x)
-    for(ulong y = 0; y < YSize; ++y)
+    for(ushort y = 0; y < YSize; ++y)
       {
 	Map[x][y] = new wsquare(this, vector2d(x, y));
 	Map[x][y]->SetWTerrain(new ocean, new atmosphere);
